% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{a.2017}
\alias{a.2017}
\title{Mangrove area}
\format{
\subsection{\code{a.2017}}{

A data frame with 1 row and 3 columns:
\describe{
\item{layer}{name of the layer}
\item{value}{raster value from which area was computed}
\item{area}{area in km2}
}
}
}
\source{
computed from mangrove.2017 using getArea
}
\usage{
a.2017
}
\description{
Area of mangroves in 2017
}
\keyword{datasets}
