% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{a}
\alias{a}
\title{Mangrove area combined}
\format{
\subsection{\code{a}}{

A data frame with 2 rows and 3 columns:
\describe{
\item{layer}{name of the layer}
\item{value}{raster value from which area was computed}
\item{area}{area in km2}
}
}
}
\source{
computed from mangrove.2000 and mangrove.2017 using getArea
}
\usage{
a
}
\description{
Area of mangroves in 2000 and 2017
}
\keyword{datasets}
