% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{AOOgrid-class}
\alias{AOOgrid-class}
\title{AOOgrid class}
\description{
A class to represent an AOO grid object
}
\section{Slots}{

\describe{
\item{\code{grid}}{an AOO grid as a shapefile}

\item{\code{AOO}}{the number of grid squares in grid}

\item{\code{params}}{a named vector of input parameters (grid size, jitter, n_jitter)}

\item{\code{pctrule}}{logical indicating whether the 1\% rule was applied}

\item{\code{input}}{the input ecosystem data used to generate the AOO grid}

\item{\code{AOOvals}}{the list of AOO values}
}}

