% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{as.list.EOO}
\alias{as.list.EOO}
\title{EOO as.list}
\usage{
\method{as.list}{EOO}(x, ...)
}
\arguments{
\item{x}{an EOO object}

\item{...}{additional arguments}
}
\description{
as.list method for EOO object
}
