% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synthesis_functions.R
\name{bundle}
\alias{bundle}
\title{Compute AOO and EOO, return as table}
\usage{
bundle(input_data, names_from = NA, ...)
}
\arguments{
\item{input_data}{Spatial object (sf or SpatRaster) of an ecosystem or species distribution.
Please use a CRS with units measured in metres.}

\item{names_from}{the name of the column containing ecosystem labels}
}
\value{
a data.frame containing AOO and EOO information for all input units as rows.
}
\description{
\code{bundle} performs AOO and EOO calculations on an input object and returns the results as a table
}
\seealso{
Other synthesis functions: 
\code{\link{list2table}()}
}
\author{
Aniko B. Toth \email{anikobtoth@gmail.com}
}
\concept{synthesis functions}
