% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change_functions.R
\name{declineForecast}
\alias{declineForecast}
\title{Decline forecasts}
\usage{
declineForecast(
  x,
  y,
  names_from_x = NA,
  names_from_y = NA,
  t1,
  year_diff,
  forecast_year = t1 + 50
)
}
\arguments{
\item{x}{SpatRaster, SpatVector, or sf object representing one or more
ecosystems or a data frame with two columns, one of them labeled "area",
and the other containing ecosystem labels names_from_x.}

\item{y}{SpatRaster, SpatVector, or sf object representing one or more
ecosystems or a data frame with two columns, one of them labeled "area",
and the other containing ecosystem labels names_from_y.}

\item{names_from_x}{name of column containing ecosystem labels. Ignored if
x is a raster.}

\item{names_from_y}{name of column containing ecosystem labels. Ignored if
y is a raster. names_from_x used if not provided.}

\item{t1}{numeric year of earliest dataset, corresponding to x}

\item{year_diff}{numeric year difference bewteen x and y inputs}

\item{forecast_year}{the desired year to which to forecast (or hindcast) change.}
}
\value{
returns a list of two elements: a table of areas and change in areas and
a table of forecasts including decline rate, forecast area, and forecast percent decline.
}
\description{
\code{declineForecast} calculates rates of change from two inputs and extrapolates the
rate of change to a desired future time interval.
}
\seealso{
Other Change functions: 
\code{\link{getArea}()},
\code{\link{getAreaChange}()},
\code{\link{getAreaTrend}()},
\code{\link{getDeclineStats}()}
}
\author{
Aniko B. Toth \email{anikobtoth@gmail.com}
}
\concept{Change functions}
