% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differenceRaster.R
\name{differenceRaster}
\alias{differenceRaster}
\title{Image Differencing.}
\usage{
differenceRaster(r1, r2)
}
\arguments{
\item{r1}{Raster object of an ecosystem or species distribution at t1}

\item{r2}{Raster object of an ecosystem or species distribution at t1}
}
\value{
Returns a raster object of the area of change between \code{r1} and
  \code{r2}.
}
\description{
Image Differencing.
}
\examples{
a = differenceRaster(raster1, raster2)
plot (raster1, col = "black")
plot (raster2, col = "darkgrey", add = TRUE)
plot(a, col = "red", add = TRUE)
}
\author{
Nicholas Murray \email{murr.nick@gmail.com}
}

