% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{EOO-class}
\alias{EOO-class}
\title{EOO class}
\description{
A class to represent an EOO convex hull object
}
\section{Slots}{

\describe{
\item{\code{pol}}{an EOO convex hull polygon}

\item{\code{EOO}}{the area of the EOO polygon as a numeric}

\item{\code{unit}}{the unit in which the area is provided}

\item{\code{input}}{the input ecosystem data used to generate the EOO polygon}
}}

