% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/futureAreaEstimate.R
\name{futureAreaEstimate}
\alias{futureAreaEstimate}
\title{Future Area Estimate}
\usage{
futureAreaEstimate(A.t1, year.t1, PRD, ARD, ARC, nYears)
}
\arguments{
\item{A.t1}{Area at time t1}

\item{year.t1}{Year of time t1}

\item{PRD}{Proportional rate of decline}

\item{ARD}{Absolute rate of decline}

\item{ARC}{Annual rate of change}

\item{nYears}{Number of years since t1 for area prediction}
}
\value{
A dataframe listing \item{area.t1}{area at time t1}
  \item{year.t1}{year of t1.} \item{prop.rate.decl}{proportional rate of
  decline.} \item{abs.rate.decl}{absolute rate of decline.}
  \item{annual.rate.change}{annual rate of change.} \item{forecast.year}{year
  for which future area estimate is made.} \item{forecast.area.prd}{future
  area as estimated with proportional rate of decline.}
  \item{forecast.area.arc}{future area as estimated with annual rate of
  change .} \item{forecast.area.ard}{future area as estimated with absolute
  rate of decline.}
}
\description{
\code{futureAreaEstimate} calculates the expected area of a distribution at a
future date using known rates of decline.
}
\examples{
a.r1 = getArea(r1) # a distribution raster
a.r2 = getArea(r2) # a distribution raster
PRD = getPRD(a.r1, a.r2, year.t1 = 1990, year.t2 = 2012)
ARD = getARD(a.r1, a.r2, year.t1 = 1990, year.t2 = 2012)
ARC = getARC(a.r1, a.r2, year.t1 = 1990, year.t2 = 2012)
area2050 <- futureAreaEstimate(A.t1 = a.r1, 2000, PRD, ARD, ARC, nYears = 50)
}
\references{
Bland, L.M., Keith, D.A., Miller, R.M., Murray, N.J. and
  Rodriguez, J.P. (eds.) 2016. Guidelines for the application of IUCN Red
  List of Ecosystems Categories and Criteria, Version 1.0. Gland,
  Switzerland: IUCN. ix + 94pp. Available at the following web site:
  \url{iucnrle.org/}
}
\seealso{
\code{\link{getArea}}, \code{\link{getPRD}}, \code{\link{getARD}},
  \code{\link{getARC}}
}
\author{
Nicholas Murray \email{murr.nick@gmail.com}
}
