% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AOO_functions.R
\name{getAOO}
\alias{getAOO}
\title{Compute Area of Occupancy (AOO)}
\usage{
getAOO(
  input.data,
  grid.size = 10000,
  names_from = NA,
  bottom.1pct.rule = TRUE,
  percent = 1,
  jitter = TRUE,
  n_jitter = 35
)
}
\arguments{
\item{input.data}{Spatial object of an ecosystem or species distribution.
Please use a CRS with units measured in metres.}

\item{grid.size}{A number specifying the width of the desired grid square (in
same units as your coordinate reference system)}

\item{names_from}{the name of the column containing ecosystem labels}

\item{bottom.1pct.rule}{Logical. If \code{TRUE}, grid cells containing the least
ecosystem area are dropped up to 1\% of the total distribution.}

\item{percent}{Numeric. The minimum percent to be applied as a threshold for
the \code{bottom.1pct.rule}}

\item{n_jitter}{the number of grids to test for ecosystems near
the AOO thresholds.}
}
\value{
an object of class AOOgrid or a list of AOOgrid objects. Ecosystems that received
an AOO of 60 cells or fewer on a first pass are run with a jittered grid with n specified by n_jitter
}
\description{
\code{getAOO} determines the number of area of occupancy (AOO) grid cells
occupied by a species or ecosystem. It includes capability for specifying
whether at least one percent of the grid cell needs to be occupied before it
is counted in the AOO. This functionality is important for assessing the IUCN
Red List of Ecosystems Criteria B.
}
\examples{
crs.UTM55S <- '+proj=utm +zone=55 +south +ellps=WGS84 +datum=WGS84 +units=m +no_defs'
r1 <- raster(ifelse((volcano<130), NA, 1), crs = crs.UTM55S)
extent(r1) <- extent(0, 6100, 0, 8700)
AOO <- getAOO(r1, 1000, bottom.1pct.rule = TRUE, percent = 1)
}
\references{
Bland, L.M., Keith, D.A., Miller, R.M., Murray, N.J. and
Rodriguez, J.P. (eds.) 2016. Guidelines for the application of IUCN Red
List of Ecosystems Categories and Criteria, Version 1.0. Gland,
Switzerland: IUCN. ix + 94pp. Available at the following web site:
\url{https://iucnrle.org/}
}
\seealso{
Other AOO functions: 
\code{\link{createGrid}()},
\code{\link{jplot}()},
\code{\link{makeAOOGrid}()},
\code{\link{top_pct}()}
}
\author{
Nicholas Murray \email{murr.nick@gmail.com}, Calvin Lee
\email{calvinkflee@gmail.com}, Aniko B. Toth \email{anikobtoth@gmail.com}
}
\concept{AOO functions}
