% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AOO_functions.R, R/getAOOSilent.R
\name{getAOOSilent}
\alias{getAOOSilent}
\alias{getAOOSilent}
\title{Alternate function for getting AOO (with custom grid)}
\usage{
getAOOSilent(ecosystem.data, grid, one.percent.rule = FALSE)

getAOOSilent(ecosystem.data, grid, one.percent.rule = FALSE)
}
\arguments{
\item{ecosystem.data}{Raster object of an ecosystem or species distribution}

\item{grid}{Custom grid to be used to calculate AOO. Usually the output of
\code{gridUncertainty}}

\item{one.percent.rule}{Logical.If \code{TRUE} one percent of the grid cell
must be occupied before it is counted in the AOO.}
}
\description{
\code{getAOOSilent} is identical to \code{getAOO}, but allows the custom
input of the grid parameter. Used for \code{gridUncertainty}.

Helper function for grid uncertainty
}
\seealso{
Other AOO functions: \code{\link{getAOO}},
  \code{\link{makeAOOGrid}}
}
\author{
Nicholas Murray \email{murr.nick@gmail.com}, Calvin Lee
  \email{calvinkflee@gmail.com}
}
