% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change_functions.R
\name{getARC}
\alias{getARC}
\title{Annual Rate of Change.}
\usage{
getARC(A.t1, A.t2, year.t1, year.t2)
}
\arguments{
\item{A.t1}{Area at time t1}

\item{A.t2}{Area at time t2}

\item{year.t1}{Year of time t1}

\item{year.t2}{Year of time t2}
}
\value{
The annual rate of change (ARC) between the two time periods
}
\description{
\code{getARC} calculates the Annual Rate of Change (ARC) given two areas and
two time points. Further information on the approach and assumptions of this
exponential decay model is available in Puyravaud (2003). Annual rate of
change is also known as instantaneous rate of change.
}
\examples{
a.r1 <- getArea(r1) # a distribution raster
a.r2 <- getArea(r2) # a distribution raster
ARC <- getARC(a.r1, a.r2, year.t1 = 1990, year.t2 = 2012)
}
\references{
Puyravaud, J.-P. 2003. Standardizing the calculation of the
annual rate of deforestation. Forest Ecology and Management, 177, 593-596.
}
\seealso{
Other Change functions: \code{\link{getARD}},
  \code{\link{getAreaChange}}, \code{\link{getArea}},
  \code{\link{getDeclineStats}}, \code{\link{getPRD}}
}
\author{
Nicholas Murray \email{murr.nick@gmail.com}, Calvin Lee
  \email{calvinkflee@gmail.com}
}
