% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change_functions.R
\name{getArea.RasterLayer}
\alias{getArea.RasterLayer}
\title{Calculates the Area of a Raster from RasterLayer.}
\usage{
\method{getArea}{RasterLayer}(x, value.to.count, ...)
}
\arguments{
\item{x}{Either a RasterLayer object. No data value should be NA}

\item{value.to.count}{Optional. Value of the cells in a RasterLayer to be
counted}

\item{...}{Addition arguments based on input format}
}
\value{
The total area of the cells of interest in km2
}
\description{
\code{getArea} reports the area of a RasterLayer object using the pixel
counting method.
}
\seealso{
Other Change functions: 
\code{\link{getArea.SpatRaster}()},
\code{\link{getArea.SpatVect}()},
\code{\link{getArea.SpatialPolygons}()},
\code{\link{getArea.sf}()},
\code{\link{getAreaLoss}()},
\code{\link{getArea}()},
\code{\link{getDeclineStats}()}
}
\author{
Nicholas Murray \email{murr.nick@gmail.com}, Calvin Lee
\email{calvinkflee@gmail.com}
}
\concept{Change functions}
