% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change_functions.R
\name{getArea.SpatRaster}
\alias{getArea.SpatRaster}
\title{Calculates the Area of a Raster from SpatRaster.}
\usage{
\method{getArea}{SpatRaster}(x, byValue, ...)
}
\arguments{
\item{x}{SpatRaster}

\item{byValue}{Logical. If TRUE, the area for each unique cell value is
returned.}

\item{...}{Addition arguments based on input format}
}
\value{
The total area of the cells of interest in km2
}
\description{
\code{getArea} reports the area of a SpatRaster object using terra::expanse
}
\seealso{
Other Change functions: 
\code{\link{getArea.RasterLayer}()},
\code{\link{getArea.SpatVect}()},
\code{\link{getArea.SpatialPolygons}()},
\code{\link{getArea.sf}()},
\code{\link{getAreaLoss}()},
\code{\link{getArea}()},
\code{\link{getDeclineStats}()}
}
\author{
Nicholas Murray \email{murr.nick@gmail.com}, Calvin Lee
\email{calvinkflee@gmail.com}
}
\concept{Change functions}
