% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change_functions.R
\name{getArea.SpatVect}
\alias{getArea.SpatVect}
\title{Calculates the Area of a Raster from SpatVect.}
\usage{
\method{getArea}{SpatVect}(x, ...)
}
\arguments{
\item{x}{A SpatVect object}

\item{...}{Addition arguments based on input format}
}
\value{
The total area of the cells of interest in km2
}
\description{
\code{getArea} reports the area of a SpatVect. object using terra::expanse
}
\seealso{
Other Change functions: 
\code{\link{getArea.RasterLayer}()},
\code{\link{getArea.SpatRaster}()},
\code{\link{getArea.SpatialPolygons}()},
\code{\link{getArea.sf}()},
\code{\link{getAreaLoss}()},
\code{\link{getArea}()},
\code{\link{getDeclineStats}()}
}
\author{
Nicholas Murray \email{murr.nick@gmail.com}, Calvin Lee
\email{calvinkflee@gmail.com}
}
\concept{Change functions}
