% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAreaChangefromRast.R
\name{getAreaChangefromRast}
\alias{getAreaChangefromRast}
\title{Area Change from  Raster.}
\usage{
getAreaChangefromRast(x, y)
}
\arguments{
\item{x}{Raster object of distribution}

\item{y}{Raster object of distribution}
}
\value{
Returns the difference in area of two raster objects in km2
}
\description{
\code{getAreaChangefromRast} reports the difference in area of two raster
objects.
}
\examples{
getAreaChangefromRast(r1, r2) # distribution rasters
}
\author{
Nicholas Murray \email{murr.nick@gmail.com}
}
\seealso{
\code{\link{getArea}} \code{\link{getAreaChangefromShp}}
\code{\link{getAreaChange}}
}

