% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAreaChangefromShp.R
\name{getAreaChangefromShp}
\alias{getAreaChangefromShp}
\title{Area Change from  Polygons.}
\usage{
getAreaChangefromShp(p1, p2)
}
\arguments{
\item{x}{Polygon of distribution}

\item{y}{Polygon of distribution}
}
\value{
Returns the difference in area of two polygon objects in km2
}
\description{
\code{getAreaChangefromShp} reports the difference in area of two polygon
datasets
}
\examples{
getAreaChangefromShp(p1, p2) # distribution rasters
}
\author{
Nicholas Murray \email{murr.nick@gmail.com}
}
\seealso{
\code{\link{getArea}} \code{\link{getAreaChangefromRast}}
\code{\link{getAreaChange}}
}

