% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EOO_functions.R
\name{getAreaEOO}
\alias{getAreaEOO}
\title{Extracts the area from an EOO polygon}
\usage{
getAreaEOO(EOO)
}
\arguments{
\item{EOO}{an object of class EOO}
}
\value{
an integer
}
\description{
\code{getEOOarea} wrapper that extracts the area slot of the EOO input
}
\seealso{
Other EOO functions: 
\code{\link{getEOO}()},
\code{\link{makeEOO}()}
}
\concept{EOO functions}
