% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCellWidth.R
\name{getCellWidth}
\alias{getCellWidth}
\title{Cell Width of Raster}
\usage{
getCellWidth(x)
}
\arguments{
\item{x}{Raster object}
}
\value{
Returns the width of each cell in a raster object
}
\description{
\code{getCellWidth} reports the width (in metres) of each raster cell.
datasets
}
\examples{
getCellWidth(x) # distribution raster
}
\author{
Nicholas Murray \email{murr.nick@gmail.com}
}
\seealso{
\code{\link{getArea}}
}

