% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EOO_functions.R
\name{getEOO}
\alias{getEOO}
\title{Calculates area of the created EOO polygon and returns a summary object with useful info.}
\usage{
getEOO(input.data, names_from = NA)
}
\arguments{
\item{input.data}{Spatial object of an ecosystem or species distribution.
Please use a CRS with units measured in metres.}

\item{names_from}{name of the column containing ecosystem names.
If missing all features will be analysed together. Only needed for vector data.}
}
\value{
An object of type EOO or a list of EOO objects that store the
EOO polygon, its area, and its input.data
}
\description{
\code{getEOO} calculates the area of the EOO polygon generated from
\code{makeEOO} the provided data and returns an EOO class object or a list of these with
defined summary and plot functions available.
}
\examples{
m <- matrix(sample(1:4, 500, replace = TRUE, prob = c(4,1,1,6)), nrow=25, ncol=20)
r1 <- terra::rast(m, crs = "EPSG:32755")
EOO <- getEOO(r1)
}
\seealso{
Other EOO functions: 
\code{\link{getAreaEOO}()},
\code{\link{makeEOO}()}
}
\author{
Nicholas Murray \email{murr.nick@gmail.com}, Calvin Lee
\email{calvinkflee@gmail.com}
}
\concept{EOO functions}
