% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change_functions.R
\name{getPercentLost}
\alias{getPercentLost}
\title{Percent change}
\usage{
getPercentLost(A.t1, A.t2)
}
\arguments{
\item{A.t1}{Area at t1}

\item{A.t2}{Area at t2}
}
\value{
The percentage change  between these two areas
}
\description{
\code{getPercentLost} reports the percentage change between two numbers (such
as areas)
}
\examples{
getAreaChange(100,80)
}
\seealso{
Other Change functions: \code{\link{getARC}},
  \code{\link{getARD}},
  \code{\link{getAreaChangefromRast}},
  \code{\link{getAreaChangefromShp}},
  \code{\link{getArea}}, \code{\link{getDeclineStats}},
  \code{\link{getPRD}}
}
\author{
Nicholas Murray \email{murr.nick@gmail.com}
}
