% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change_functions.R
\name{getPRD}
\alias{getPRD}
\title{Absolute Rate of Decline.}
\usage{
getPRD(A.t1, A.t2, year.t1, year.t2)
}
\arguments{
\item{A.t1}{Area at time t1}

\item{A.t2}{Area at time t2}

\item{year.t1}{Year of time t1}

\item{year.t2}{Year of time t2}
}
\value{
Returns the proportional rate of decline (PRD) between the two time
periods
}
\description{
\code{getPRD} calculates the Proportional Rate of Decline (PRD) given two
areas and two time points. Further information on the approach and
assumptions of this model is available in Keith et al (2003). PRD is
sometimes known as trajectory (eg. Waycott et al 2009).
}
\examples{
a.r1 = getArea(r1) # a distribution raster
a.r2 = getArea(r2) # a distribution raster
PRD = getPRD(a.r1, a.r2, year.t1 = 1990, year.t2 = 2012)
}
\references{
Keith, D. A., Orscheg, C., Simpson, C. C., Clarke, P. J., Hughes,
L., Kennelly, S. J., Major, R. E., Soderquist, T. R., Wilson, A. L. &
Bedward, M. 2009. A new approach and case study for estimating extent and
rates of habitat loss for ecological communities. Biological Conservation,
142, 1469-1479.

Bland, L.M., Keith, D.A., Miller, R.M., Murray, N.J. and
Rodriguez, J.P. (eds.) 2016. Guidelines for the application of IUCN Red
List of Ecosystems Categories and Criteria, Version 1.0. Gland, Switzerland:
IUCN. ix + 94pp. Available at the following web site: \url{iucnrle.org/}
}
\seealso{
Other Change functions: \code{\link{getARC}},
  \code{\link{getARD}},
  \code{\link{getAreaChangefromRast}},
  \code{\link{getAreaChangefromShp}},
  \code{\link{getArea}}, \code{\link{getDeclineStats}},
  \code{\link{getPercentLost}}
}
\author{
Nicholas Murray \email{murr.nick@gmail.com}
}
