% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gridUncertainty.R
\name{gridUncertainty}
\alias{gridUncertainty}
\title{Compute Area of Occupancy (AOO) accounting for grid uncertainty}
\usage{
gridUncertainty(x, grid.rast, nSim = 100)
}
\arguments{
\item{x}{Raster object of an ecosystem or species distribution}

\item{nSim}{specifies the number of times to repeat the AOO calculation
with randomly sampled grid origin coordinates}

\item{grid}{Raster object of a regular grid suitable for assessing AOO}
}
\value{
A list of dataframes raw.sims and summary.sims
}
\description{
\code{gridUncertainty} determines the number of area of occupancy (AOO) grid
cells occupied by a species or ecosystem. The key difference with
\code{getAOO} is that it accounts for the impact of the location of the AOO
grid by randomly moving the grid and recomputing AOO at each step.
}
\author{
Nicholas Murray \email{murr.nick@gmail.com}
}
\references{
Bland, L.M., Keith, D.A., Miller, R.M., Murray, N.J. and
Rodriguez, J.P. (eds.) 2016. Guidelines for the application of IUCN Red
List of Ecosystems Categories and Criteria, Version 1.0. Gland, Switzerland:
IUCN. ix + 94pp. Available at the following web site: \url{iucnrle.org/}
}
\seealso{
\code{\link{createGrid}} \code{\link{getAOO}}
}

