% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AOO_functions.R
\name{jplot}
\alias{jplot}
\title{Make elbow plot to check jitter iterations}
\usage{
jplot(x)
}
\arguments{
\item{x}{an AOOgrid object}
}
\value{
NULL; plots min AOO against number of reps
}
\description{
\code{jplot} creates an elbow plot of the min AOO against
the number of grid replicates run. Jplots that fall to the
minimum value well before the highest n are robust.
}
\seealso{
Other AOO functions: 
\code{\link{createGrid}()},
\code{\link{getAOO}()},
\code{\link{makeAOOGrid}()},
\code{\link{top_pct}()}
}
\author{
Aniko B. Toth \email{anikobtoth@gmail.com}
}
\concept{AOO functions}
