% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synthesis_functions.R
\name{list2table}
\alias{list2table}
\title{Summarises a list of EOO or AOOgrid objects in a table}
\usage{
list2table(l)
}
\arguments{
\item{l}{a list of EOO or AOOgrid objects}
}
\value{
a data.frame
}
\description{
\code{list2table} wrapper that summarises a list of EOO or AOOgrid objects as a table.
}
\seealso{
Other synthesis functions: 
\code{\link{bundle}()}
}
\concept{synthesis functions}
