% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AOO_functions.R
\name{makeAOOGrid}
\alias{makeAOOGrid}
\title{Create Area of Occupancy (AOO) grid for an ecosystem or species distribution}
\usage{
makeAOOGrid(
  input.data,
  grid.size = 10000,
  names_from = NA,
  bottom.1pct.rule = TRUE,
  percent = 1,
  jitter = TRUE,
  n_jitter = 35
)
}
\arguments{
\item{input.data}{Spatial object (sf or SpatRaster) of an ecosystem or species distribution.
Please use a CRS with units measured in metres.}

\item{grid.size}{A number specifying the width of the desired grid square (in
same units as your coordinate reference system)}

\item{names_from}{the name of the column containing ecosystem labels}

\item{bottom.1pct.rule}{Logical. If \code{TRUE}, grid cells containing the least
ecosystem area are dropped up to 1\% of the total distribution.}

\item{percent}{Numeric. The minimum percent to be applied as a threshold for
the \code{bottom.1pct.rule}}

\item{jitter}{logical. Whether grid randomization should be applied to units with low grid counts.}

\item{n_jitter}{the number of grids to test for ecosystems near
the AOO thresholds. Ignored if jitter = FALSE.}
}
\value{
A shapefile of grid cells occupied by an ecosystem or species, or a
list of these if multiple ecosystems were input.
}
\description{
\code{makeAOOGrid} is a generic function that creates grids representing the
area of occupancy for distributions based on the input spatial data. It
includes capability for specifying whether the least occupied cells collectively
containing less than 1\% of the ecosystem are counted in the AOO. This functionality is
important for assessing the IUCN Red List of Ecosystems Criteria B.
}
\examples{
m <- matrix(sample(1:4, 500, replace = TRUE, prob = c(4,1,1,6)), nrow=25, ncol=20)
r1 <- terra::rast(m, crs = "EPSG:32755")
AOO_grid <- makeAOOGrid(r1, grid.size = 3)
}
\references{
Bland, L.M., Keith, D.A., Miller, R.M., Murray, N.J. and
Rodriguez, J.P. (eds.) 2016. Guidelines for the application of IUCN Red
List of Ecosystems Categories and Criteria, Version 1.0. Gland,
Switzerland: IUCN. ix + 94pp. Available at the following web site:
\url{https://iucnrle.org/}
}
\seealso{
Other AOO functions: 
\code{\link{createGrid}()},
\code{\link{getAOO}()},
\code{\link{jplot}()},
\code{\link{top_pct}()}
}
\author{
Nicholas Murray \email{murr.nick@gmail.com}, Calvin Lee
\email{calvinkflee@gmail.com}, Aniko B. Toth \email{anikobtoth@gmail.com}
}
\concept{AOO functions}
