% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{plot,EOO,ANY-method}
\alias{plot,EOO,ANY-method}
\title{EOO plot}
\usage{
\S4method{plot}{EOO,ANY}(x, title = x@name)
}
\arguments{
\item{x}{an EOO object}

\item{title}{Plot title, defaults to name of assessment unit.}

\item{...}{Additional graphical parameters passed to \code{\link[graphics]{plot}}.}
}
\description{
plot method for EOO object
}
