% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{plot,trend,ANY-method}
\alias{plot,trend,ANY-method}
\title{trend plot}
\usage{
\S4method{plot}{trend,ANY}(x, y, ...)
}
\arguments{
\item{x}{an trend object}

\item{y}{Ignored. Included for consistency with the \code{plot} generic.}

\item{...}{Additional graphical parameters passed to \code{\link[graphics]{plot}}.}
}
\description{
plot method for trend object
}
