% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redListPalette.R
\name{redListPalette}
\alias{redListPalette}
\title{Red List Palette.}
\usage{
redListPalette()
}
\value{
Returns a character vector of the colours used for each category of
  threat by the IUCN
}
\description{
\code{redListPalette} creates a character vector of colours suggested by the
IUCN for listing the 8 categories of threat
}
\examples{
pie(rep(1,8), col=redListPalette(), main = "Red List Palette")
n <- colorRampPalette(redListPalette())(50)
pie(rep(1,50), col=n, frame = F)
}
\author{
Nicholas Murray \email{murr.nick@gmail.com}
}
