% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{show,AOOgrid-method}
\alias{show,AOOgrid-method}
\title{AOOgrid show}
\usage{
\S4method{show}{AOOgrid}(object)
}
\arguments{
\item{object}{an AOOgrid object}
}
\description{
show method for AOOgrid object
}
