% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{summary,AOOgrid-method}
\alias{summary,AOOgrid-method}
\title{AOOgrid summary}
\usage{
\S4method{summary}{AOOgrid}(object)
}
\arguments{
\item{object}{an AOOgrid object}
}
\description{
summary method for AOOgrid object
}
