% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{summary,trend-method}
\alias{summary,trend-method}
\title{trend summary}
\usage{
\S4method{summary}{trend}(object)
}
\arguments{
\item{object}{a trend object}
}
\description{
summary method for trend object
}
