% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AOO_functions.R
\name{top_pct}
\alias{top_pct}
\title{Identify positions of the bottom 1\% of ecosystem area in AOO grid}
\usage{
top_pct(v, pct = 99)
}
\arguments{
\item{v}{A numeric vector.}

\item{pct}{percent of area to drop}
}
\value{
a numeric vector indicating the indeces of the elements to keep.
}
\description{
\code{top_pct} returns the vector positions of the largest elements
collectively comprising a given percentage more of the vector sum. This function
helps perform the bottom.1pct.rule when selecting the AOO grid by
identifying grid positions to keep.
}
\references{
IUCN (2024). Guidelines for the application of IUCN Red
List of Ecosystems Categories and Criteria, Version 2.0. Keith, D.A.,
Ferrer-Paris, J.R., Ghoraba, S.M.M., Henriksen, S., Monyeki, M., Murray,
N.J., Nicholson, E., Rowland, J., Skowno, A., Slingsby, J.A., Storeng,
A.B., Valderrábano, M. & Zager, I. (Eds.). Gland, Switzerland: IUCN.
ix + 94pp. Available at the following web site:
\url{https://iucnrle.org/}
}
\seealso{
Other AOO functions: 
\code{\link{createGrid}()},
\code{\link{getAOO}()},
\code{\link{jplot}()},
\code{\link{makeAOOGrid}()}
}
\author{
Aniko B. Toth \email{anikobtoth@gmail.com}
}
\concept{AOO functions}
