% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{trend-class}
\alias{trend-class}
\title{trend class}
\description{
A class to represent change in area over time for an ecosystem.
}
\section{Slots}{

\describe{
\item{\code{input}}{binary raster or list of sf POLYGON objects representing the input ecosystem extent}

\item{\code{areas}}{the area of the ecosystem in each layer or list element}

\item{\code{model}}{a fitted model object typically of the class lm containing the model used to calculate the trend}

\item{\code{netdiff}}{difference between the area in the first layer or list element and the last.}

\item{\code{diff}}{raster stack showing change over layers or list elements of the ecosystem extent.}
}}

