% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change_functions.R
\name{trend_fitting_functions}
\alias{trend_fitting_functions}
\alias{fit_lm_ll}
\alias{fit_glm_ll}
\alias{fit_spline}
\title{Fit linear model with log-link.}
\usage{
fit_lm_ll(df)

fit_glm_ll(df)

fit_spline(df)
}
\arguments{
\item{df}{a data frame containing the columns "area" and "t" for time. Time
can be an index or year.}
}
\value{
an lm object
}
\description{
\code{fit_lm_ll} produces a simple model to estimate trends in short time series.
}
\author{
Aniko B. Toth \email{anikobtoth@gmail.com}
}
