% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createGrid.R
\name{createGrid}
\alias{createGrid}
\title{Create Area of Occupancy (AOO) Grid.}
\usage{
createGrid(x, grid.size)
}
\arguments{
\item{x}{Raster object of an ecosystem or species distribution}

\item{grid.size}{A number specifying the width of the desired grid square
(in metres)}
}
\value{
A regular grid raster with extent \code{ecosystem.data} and grid size
  \code{grid.size}. Each grid square has a unique identification number
}
\description{
Create Area of Occupancy (AOO) Grid.
}
\examples{
createGrid(raster1, 10000) # a 10-km grid with extent of raster1
createGrid(raster1, 2000) # a 2-km grid with extent of raster1
}
\references{
Bland, L.M., Keith, D.A., Miller, R.M., Murray, N.J. and
Rodriguez, J.P. (eds.) 2016. Guidelines for the application of IUCN Red
List of Ecosystems Categories and Criteria, Version 1.0. Gland, Switzerland:
IUCN. ix + 94pp. Available at the following web site: \url{iucnrle.org/}
}
\seealso{
\code{\link{getAOO}}, \code{\link{getRLEReport}}
}
\author{
Nicholas Murray \email{murr.nick@gmail.com}
}
