% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAOO.R
\name{getAOO}
\alias{getAOO}
\title{Compute Area of Occupancy (AOO)}
\usage{
getAOO(ecosystem.data, grid, one.percent.rule = TRUE)
}
\arguments{
\item{grid}{Raster object of a regular grid suitable for assessing AOO}

\item{x}{Raster object of an ecosystem or species distribution}
}
\value{
The number of grid cells occupied by the ecosystem or species
}
\description{
\code{getAOO} determines the number of area of occupancy (AOO) grid cells
occupied by a species or ecosystem. It includes capability for specifying
the proportion of each grid cell that needs to be occupied before the grid
cell is counted in the AOO. This functionality is important for assessing
the IUCN Red List of Ecosystems Criteria B.
}
\examples{
n = createGrid(r1, resolution = 10000)
plot (n)
plot (r1, add = TRUE)
AOO = getAOO(r1, n, one.percent.rule = F)
AOO # number of grid cells occupied by an ecosystem or species
}
\author{
Nicholas Murray \email{murr.nick@gmail.com}
}
\references{
Bland, L.M., Keith, D.A., Miller, R.M., Murray, N.J. and
Rodríguez, J.P. (eds.) 2016. Guidelines for the application of IUCN Red
List of Ecosystems Categories and Criteria, Version 1.0. Gland, Switzerland:
IUCN. ix + 94pp. Available at the following web site: \url{iucnrle.org/}
}
\seealso{
\code{\link{createGrid}}
}

