% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AOO_functions.R
\name{getAOO}
\alias{getAOO}
\title{Compute Area of Occupancy (AOO)}
\usage{
getAOO(ecosystem.data, grid.size, one.percent.rule = TRUE)
}
\arguments{
\item{ecosystem.data}{Raster object of an ecosystem or species distribution.
Please use a CRS with units measured in metres.}

\item{grid.size}{A number specifying the width of the desired grid square (in
same units as your coordiante reference system)}

\item{one.percent.rule}{Logical.If \code{TRUE} one percent of the grid cell
must be occupied before it is counted in the AOO.}
}
\value{
The number of grid cells occupied by the ecosystem or species
}
\description{
\code{getAOO} determines the number of area of occupancy (AOO) grid cells
occupied by a species or ecosystem. It includes capability for specifying
whether at least one percent of the grid cell needs to be occupied before it
is counted in the AOO. This functionality is important for assessing the IUCN
Red List of Ecosystems Criteria B.
}
\examples{
n = createGrid(r1, 10000)
plot (n)
plot (r1, add = TRUE)
AOO = getAOO(r1, n, one.percent.rule = F)
AOO # number of grid cells occupied by an ecosystem or species
}
\references{
Bland, L.M., Keith, D.A., Miller, R.M., Murray, N.J. and
  Rodriguez, J.P. (eds.) 2016. Guidelines for the application of IUCN Red
  List of Ecosystems Categories and Criteria, Version 1.0. Gland,
  Switzerland: IUCN. ix + 94pp. Available at the following web site:
  \url{iucnrle.org/}
}
\seealso{
Other AOO functions: \code{\link{getAOOSilent}},
  \code{\link{makeAOOGrid}}
}
\author{
Nicholas Murray \email{murr.nick@gmail.com}, Calvin Lee
  \email{calvinkflee@gmail.com}
}
