% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAOOSilent.R
\name{getAOOSilent}
\alias{getAOOSilent}
\title{Helper function for grid uncertainty}
\usage{
getAOOSilent(ecosystem.data, grid, one.percent.rule = FALSE)
}
\description{
Helper function for grid uncertainty
}
