% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change_functions.R
\name{getArea}
\alias{getArea}
\title{Calculates the Area of ecosystems in spatial data}
\usage{
getArea(x, names_from = NA, ...)
}
\arguments{
\item{x}{A SpatRaster, SpatVector, or an sf object with POLYGONS geometry.}

\item{names_from}{a column names containing ecosystem labels, as a string or
dplyr-style column name. Only required for SpatVector and sf types. Units are
assumed to be delineated by raster value for SpatRasters.}

\item{...}{Addition arguments based on input format}
}
\value{
A data frame containing ecosystem identifiers and the total area of the
ecosystem units in x as a units vector (km^2). For raster bricks it also contains
the layer number.
}
\description{
\code{getArea} reports the area of ecosystem units provided as spatial data
}
\examples{
m <- matrix(sample(1:4, 500, replace = TRUE, prob = c(4,1,1,6)), nrow=25, ncol=20)
r1 <- terra::rast(m, crs = "EPSG:32755")
a.r1 <- getArea(r1) # area of all non-NA cells in r1
}
\seealso{
Other Change functions: 
\code{\link{getAreaChange}()},
\code{\link{getAreaTrend}()},
\code{\link{getDeclineStats}()}
}
\author{
Nicholas Murray \email{murr.nick@gmail.com}, Calvin Lee
\email{calvinkflee@gmail.com}, Aniko B. Toth \email{anikobtoth@gmail.com}
}
\concept{Change functions}
