% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getArea.R
\name{getArea}
\alias{getArea}
\title{Area of Raster.}
\usage{
getArea(x)
}
\arguments{
\item{x}{A binary raster object with cells containing no data identified as
NA}
}
\value{
Returns the area of a raster object in km2
}
\description{
\code{getArea} reports the area of a raster object using the pixel counting
method. Requires a binary raster as input.
}
\examples{
a.r1 = getArea(r1) # a distribution raster
}
\seealso{
\code{\link{getCellRes}}
}
\author{
Nicholas Murray \email{murr.nick@gmail.com}
}
