% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change_functions.R
\name{getArea}
\alias{getArea}
\title{Calculates the Area of a Raster.}
\usage{
getArea(x, value.to.count)
}
\arguments{
\item{x}{A raster object with cells containing no data identified as NA}

\item{value.to.count}{Optional. Value of the cells to be counted}
}
\value{
The area of the cells of interest
}
\description{
\code{getArea} reports the area of a raster object using the pixel counting
method.
}
\examples{
a.r1 <- getArea(r1) # area of all non-NA cells in r1
a.r1 <- getArea(r1, 1) # area of cells labelled 1 in r1
}
\seealso{
Other Change functions: \code{\link{getARC}},
  \code{\link{getARD}}, \code{\link{getAreaChange}},
  \code{\link{getDeclineStats}}, \code{\link{getPRD}}
}
\author{
Nicholas Murray \email{murr.nick@gmail.com}, Calvin Lee
  \email{calvinkflee@gmail.com}
}
