% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change_functions.R
\name{getAreaChange}
\alias{getAreaChange}
\title{Area change between two inputs in km2}
\usage{
getAreaChange(x, y, names_from_x = NA, names_from_y = NA)
}
\arguments{
\item{x}{SpatRaster, SpatVector, or sf object representing one or more
ecosystems or a data frame with two columns, one of them labeled "area",
and the other containing ecosystem labels names_from_x.}

\item{y}{SpatRaster, SpatVector, or sf object representing one or more
ecosystems or a data frame with two columns, one of them labeled "area",
and the other containing ecosystem labels names_from_y.}

\item{names_from_x}{name of column containing ecosystem labels. Ignored if
x is a raster.}

\item{names_from_y}{name of column containing ecosystem labels. Ignored if
y is a raster. names_from_x used if not provided.}
}
\value{
Returns a table containing ecosystem labels, areas, and the
difference in area of the two inputs in km2
}
\description{
\code{getAreaChange} reports the difference in area between two inputs. Inputs
can be  SpatRaster, SpatVector, sf or a data frame of areas and may contain
data for multiple ecosystem types. Ensure x and y are the same data type.
If using data frame as input, ensure areas are measured in km2
}
\examples{
crs.UTM55S <- '+proj=utm +zone=55 +south +ellps=WGS84 +datum=WGS84 +units=m +no_defs'
r1 <- raster(ifelse((volcano<130), NA, 1), crs = crs.UTM55S)
extent(r1) <- extent(0, 6100, 0, 8700)
r2 <- raster(ifelse((volcano<145), NA, 1), crs = crs.UTM55S)
extent(r2) <- extent(0, 6100, 0, 8700)
a.dif <- getAreaChange(r1, r2) # distribution rasters
}
\seealso{
Other Change functions: 
\code{\link{getArea}()},
\code{\link{getAreaTrend}()},
\code{\link{getDeclineStats}()}
}
\author{
Nicholas Murray \email{murr.nick@gmail.com}, Calvin Lee
\email{calvinkflee@gmail.com}, Aniko B. Toth \email{anikobtoth@gmail.com}
}
\concept{Change functions}
