% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAreaChange.R
\name{getAreaChange}
\alias{getAreaChange}
\title{Area Change.}
\usage{
getAreaChange(A.t1, A.t2)
}
\arguments{
\item{A.t1}{Area at t1}

\item{A.t2}{Area at t2}
}
\value{
Returns the difference between two areas.
}
\description{
\code{getAreaChange} reports the difference two areas.
}
\examples{
getAreaChange(100,80)
}
\seealso{
\code{\link{getAreaChangefromRast}}
\code{\link{getAreaChangefromShp}}
}
\author{
Nicholas Murray \email{murr.nick@gmail.com}
}
