% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change_functions.R
\name{getAreaChangefromRast}
\alias{getAreaChangefromRast}
\title{Area Change from Raster.}
\usage{
getAreaChangefromRast(x, y)
}
\arguments{
\item{x}{Raster object of distribution}

\item{y}{Raster object of distribution}
}
\value{
Returns the difference in area of two raster objects in km2
}
\description{
\code{getAreaChangefromRast} reports the difference in area of two raster
objects.
}
\examples{
getAreaChangefromRast(r1, r2) # distribution rasters
}
\seealso{
Other Change functions: \code{\link{getARC}},
  \code{\link{getARD}}, \code{\link{getAreaChangefromShp}},
  \code{\link{getArea}}, \code{\link{getDeclineStats}},
  \code{\link{getPRD}}, \code{\link{getPercentLost}}
}
\author{
Nicholas Murray \email{murr.nick@gmail.com}
}
