% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change_functions.R
\name{getAreaChangefromShp}
\alias{getAreaChangefromShp}
\title{Area Change from Polygons.}
\usage{
getAreaChangefromShp(p1, p2)
}
\arguments{
\item{x}{Polygon of distribution}

\item{y}{Polygon of distribution}
}
\value{
Returns the difference in area of two polygon objects in km2
}
\description{
\code{getAreaChangefromShp} reports the difference in area of two polygon
datasets
}
\examples{
getAreaChangefromShp(p1, p2) # distribution rasters
}
\seealso{
Other Change functions: \code{\link{getARC}},
  \code{\link{getARD}},
  \code{\link{getAreaChangefromRast}},
  \code{\link{getArea}}, \code{\link{getDeclineStats}},
  \code{\link{getPRD}}, \code{\link{getPercentLost}}
}
\author{
Nicholas Murray \email{murr.nick@gmail.com}
}
