% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EOO_functions.R
\name{getAreaEOO}
\alias{getAreaEOO}
\title{Calculates area of the created EOO polygon.}
\usage{
getAreaEOO(EOO.polygon)
}
\arguments{
\item{EOO.polygon}{An object of class SpatialPolygons, usually the output
from \code{makeEOO}.}
}
\value{
The area of the \code{EOO.polygon} in km2
}
\description{
\code{getAreaEOO} calculates the area of the EOO polygon generated from
\code{makeEOO} the provided data
}
\seealso{
Other EOO functions: \code{\link{makeEOO}}
}
\author{
Nicholas Murray \email{murr.nick@gmail.com}, Calvin Lee
  \email{calvinkflee@gmail.com}
}
