% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change_functions.R
\name{getAreaTrend}
\alias{getAreaTrend}
\title{Calculate area trends of ecosystems}
\usage{
getAreaTrend(x, names_from = NA)
}
\arguments{
\item{x}{SpatRaster with multiple layers or data frame containing two numerical
columns "area" and "time". The data.frame can contain an additional column containing
a key, whose name should be placed in names_from.}

\item{names_from}{name of column containing ecosystem labels. Ignored if
x is a raster.}
}
\value{
returns a list
}
\description{
\code{getAreaTrend} is used to calculate changes in area over time. Output
is a list can be used to easily visualise trends for one ecosystem at a time.
}
\examples{
crs.UTM55S <- '+proj=utm +zone=55 +south +ellps=WGS84 +datum=WGS84 +units=m +no_defs'
r1 <- raster(ifelse((volcano<130), NA, 1), crs = crs.UTM55S)
extent(r1) <- extent(0, 6100, 0, 8700)
r2 <- raster(ifelse((volcano<145), NA, 1), crs = crs.UTM55S)
extent(r2) <- extent(0, 6100, 0, 8700)
a.dif <- getAreaChange(r1, r2) # distribution rasters
}
\seealso{
Other Change functions: 
\code{\link{getArea}()},
\code{\link{getAreaChange}()},
\code{\link{getDeclineStats}()}
}
\author{
Aniko B. Toth \email{anikobtoth@gmail.com}
}
\concept{Change functions}
