% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EOO_functions.R
\name{getEOO}
\alias{getEOO}
\title{Calculates area of the created EOO polygon.}
\usage{
getEOO(input.data, names_from = NA)
}
\arguments{
\item{EOO.polygon}{An object of class SpatVect, usually the output
from \code{makeEOO}.}

\item{unit}{Character. Output unit of area. One of "m", "km", or "ha"}
}
\value{
An object of type EOO or a list of EOO objects that store the
EOO polygon, its area, and its input.data
}
\description{
\code{getEOO} calculates the area of the EOO polygon generated from
\code{makeEOO} the provided data and returns a formatted output with defined summary and plot functions
}
\examples{
library(terra)
crs.UTM55S <- '+proj=utm +zone=55 +south +ellps=WGS84 +datum=WGS84 +units=m +no_defs'
r1 <- rast(ifelse((volcano<130), NA, 1), crs = crs.UTM55S)
ext(r1) <- c(0, 6100, 0, 8700)
EOO.polygon <- makeEOO(r1)
EOO.area <- getAreaEOO(EOO.polygon)
}
\seealso{
Other EOO functions: 
\code{\link{makeEOO}()}
}
\author{
Nicholas Murray \email{murr.nick@gmail.com}, Calvin Lee
\email{calvinkflee@gmail.com}
}
\concept{EOO functions}
