% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPercentLost.R
\name{getPercentLost}
\alias{getPercentLost}
\title{Percent change}
\usage{
getPercentLost(A.t1, A.t2)
}
\arguments{
\item{A.t1}{Area at t1}

\item{A.t2}{Area at t2}
}
\value{
Returns the percentage change  between these two areas
}
\description{
\code{getPercentLost} reports the percentage change between two numbers (such
as areas)
}
\examples{
getAreaChange(100,80)
}
\author{
Nicholas Murray \email{murr.nick@gmail.com}
}
\seealso{
\code{\link{getAreaChangefromRast}}
\code{\link{getAreaChangefromShp}} \code{\link{getAreaChange}}
}

