% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gridUncertainty.R
\name{gridUncertainty}
\alias{gridUncertainty}
\title{Compute Area of Occupancy (AOO) accounting for grid uncertainty}
\usage{
gridUncertainty(ecosystem.data, grid.size, split, one.percent.rule = TRUE,
  percent)
}
\arguments{
\item{split}{Specifies the number of ways to split the grid in ONE axis.}

\item{percent}{Numeric. The minimum percent to be applied as a threshold for
the \code{min.percent.rule}}

\item{x}{Raster object of an ecosystem or species distribution}

\item{grid}{Raster object of a regular grid suitable for assessing AOO}

\item{min.percent.rule}{Logical. If \code{TRUE}, a minimum area threshold
must be passed before a grid is counted as an AOO grid.}
}
\value{
List containing vector of length split*split of calculated AOO for
  each shift position, a list of summary statistics for the vector, a list of
  grids which generated the smallest AOO as RasterLayers, and the movements
  which resulted in the grids with the smallest AOO.
}
\description{
\code{gridUncertainty} determines the number of area of occupancy (AOO) grid
cells occupied by a species or ecosystem. The key difference with
\code{getAOO} is that it accounts for the impact of the location of the AOO
grid by systematically moving the grid and recomputing AOO at each step.
}
\examples{
x <- gridUncertainty(r1, 10000, 3, F)
x$AOO.list # vector of length 9 of calculated AOO for each grid position
x$stats # dataframe returning the number of shifts performed, summary statistics
for the calculated AOOs, and the shift values which created the grid
position(s) with the smallest AOO.
x$min.AOO.grid.list # list of RasterLayers of the grids which all return the
minimum AOO.
x$min.AOO.shifts # dataframe containing information of the index of the shift
which created the grids with the smallest AOO and the x and y shifts for them.
}
\references{
Bland, L.M., Keith, D.A., Miller, R.M., Murray, N.J. and
  Rodriguez, J.P. (eds.) 2016. Guidelines for the application of IUCN Red
  List of Ecosystems Categories and Criteria, Version 1.0. Gland,
  Switzerland: IUCN. ix + 94pp. Available at the following web site:
  \url{iucnrle.org/}
}
\seealso{
\code{\link{createGrid}} \code{\link{getAOO}}
  \code{\link{getAOOSilent}}
}
\author{
Nicholas Murray \email{murr.nick@gmail.com}, Calvin Lee
  \email{calvinkflee@gmail.com}
}
