% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gridUncertainty_functions.R
\name{gridUncertainty}
\alias{gridUncertainty}
\title{Function to compute AOO with grid uncertainty systematically with stopping rule}
\usage{
gridUncertainty(ecosystem.data, grid.size, n.AOO.improvement,
  min.percent.rule = T, percent = 1)
}
\arguments{
\item{ecosystem.data}{Raster object of an ecosystem or species distribution.
Please use a CRS with units measured in metres.}

\item{grid.size}{A number specifying the width of the desired grid square (in
same units as your coordiante reference system)}

\item{n.AOO.improvement}{Specifies the minimum number of rounds the
calculated AOO is not improved before stopping the function.}

\item{min.percent.rule}{Logical. If \code{TRUE}, a minimum area threshold
must be passed before a grid is counted as an AOO grid.}

\item{percent}{Numeric. The minimum percent to be applied as a threshold for
the \code{min.percent.rule}}
}
\value{
A list containing the following:
\itemize{
 \item Data frame of results showing the minimum AOO calculated for each
 shift scenario
 \item List of rasterLayers which were used to generate the
 minimum AOO specified for each shift scenario
}
}
\description{
\code{gridUncertainty} determines the number of area of occupancy (AOO) grid
cells occupied by a species or ecosystem systematically. It will only stop
when the AOO calculated does not improve (decrease) after a set number of
split scenarios.
}
\examples{
AOO.df <- gridUncertainty(r, 10000, 4, T, 1)
AOO <- min(AOO.df$min.AOO)
}
\references{
Bland, L.M., Keith, D.A., Miller, R.M., Murray, N.J. and
  Rodriguez, J.P. (eds.) 2016. Guidelines for the application of IUCN Red
  List of Ecosystems Categories and Criteria, Version 1.0. Gland,
  Switzerland: IUCN. ix + 94pp. Available at the following web site:
  \url{iucnrle.org/}
}
\seealso{
Other gridUncertainty functions: \code{\link{gridUncertaintyBase}},
  \code{\link{gridUncertaintyRandomManual}},
  \code{\link{gridUncertaintySimulation}}
}
\author{
Calvin Lee \email{calvinkflee@gmail.com}
}
