% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gridUncertainty_functions.R
\name{gridUncertaintyBase}
\alias{gridUncertaintyBase}
\title{Base function to compute AOO with grid uncertainty}
\usage{
gridUncertaintyBase(ecosystem.data, grid.size, splits,
  min.percent.rule = TRUE, percent = 1)
}
\arguments{
\item{ecosystem.data}{Raster object of an ecosystem or species distribution.
Please use a CRS with units measured in metres.}

\item{grid.size}{A number specifying the width of the desired grid square (in
same units as your coordiante reference system)}

\item{min.percent.rule}{Logical. If \code{TRUE}, a minimum area threshold
must be passed before a grid is counted as an AOO grid.}

\item{percent}{Numeric. The minimum percent to be applied as a threshold for
the \code{min.percent.rule}.}

\item{split}{Specifies the number of ways to split the grid in ONE axis.}
}
\value{
List containing the following:
\itemize{
 \item Vector of length split*split of calculated AOO for each shifted grid
 \item Data frame of summary statistics for the results
 \item Data frame showing the distance shifted in x and y directions used to
 create the AOO grid(s) which return the smallest AOO
 \item List of RasterLayer(s) containing the AOO grid(s) which return the
 smallest AOO
}
}
\description{
\code{gridUncertaintyBase} helps determine the minimum number of area of
occupancy (AOO) grid cells occupied by a species or ecosystem. It varies the
location of the AOO grid by shifting in systematically in both x- and y-
axes, returning summary statistics for the range of AOOs calculated, and the
RasterLayer(s) containing the grids with the minimum AOO. It is the base
function which is used by \code{gridUncertainty} and
\code{gridUncertaintySimulation}
}
\examples{
x <- gridUncertaintyBase(r1, 10000, 3, F)
x$AOO.list # vector of length 9 of calculated AOO for each grid position
x$stats # dataframe returning the number of shifts performed, summary statistics
for the calculated AOOs, and the shift values which created the grid
position(s) with the smallest AOO.
x$min.AOO.grid.list # list of RasterLayers of the grids which all return the
minimum AOO.
x$min.AOO.shifts # dataframe containing information of the index of the shift
which created the grids with the smallest AOO and the x and y shifts for them.
}
\references{
Bland, L.M., Keith, D.A., Miller, R.M., Murray, N.J. and
  Rodriguez, J.P. (eds.) 2016. Guidelines for the application of IUCN Red
  List of Ecosystems Categories and Criteria, Version 1.0. Gland,
  Switzerland: IUCN. ix + 94pp. Available at the following web site:
  \url{iucnrle.org/}
}
\seealso{
\code{\link{createGrid}} \code{\link{getAOO}}
  \code{\link{getAOOSilent}}

Other gridUncertainty functions: \code{\link{gridUncertaintySimulation}},
  \code{\link{gridUncertainty}}
}
\author{
Nicholas Murray \email{murr.nick@gmail.com}, Calvin Lee
  \email{calvinkflee@gmail.com}
}
