% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gridUncertainty_functions.R
\name{gridUncertaintySimulation}
\alias{gridUncertaintySimulation}
\title{Function to investigate behaviour of AOO under various splits}
\usage{
gridUncertaintySimulation(ecosystem.data, grid.size, simulations,
  min.percent.rule, percent)
}
\arguments{
\item{ecosystem.data}{Raster object of an ecosystem or species distribution.
Please use a CRS with units measured in metres.}

\item{grid.size}{A number specifying the width of the desired grid square (in
same units as your coordiante reference system)}

\item{simulations}{Specifies the maximum number of splits to be performed on
the generated grid}

\item{min.percent.rule}{Logical. If \code{TRUE}, a minimum area threshold
must be passed before a grid is counted as an AOO grid.}

\item{percent}{Numeric. The minimum percent to be applied as a threshold for
the \code{min.percent.rule}}
}
\value{
Data frame of results showing the minimum and maximum AOO calculated
  for each grid shift scenario
}
\description{
\code{gridUncertaintySimulation} returns the number of area of occupancy
(AOO) grid cells occupied by a species or ecosystem in incremental splits
using \code{gridUncertaintyBase}.
}
\examples{
AOO.df <- gridUncertaintySimulation(r, 10000, 10, T, 1)
plot(AOO.df$n.splits, AOO.df$min.AOO,
     type = 'l', col = 'red', ylim = c(0, 10), xlab = 'Number of splits', ylab = 'Minimum AOO')
lines(AOO.df$n.splits, AOO.df$max.AOO, col = 'blue')
}
\references{
Bland, L.M., Keith, D.A., Miller, R.M., Murray, N.J. and
  Rodriguez, J.P. (eds.) 2016. Guidelines for the application of IUCN Red
  List of Ecosystems Categories and Criteria, Version 1.0. Gland,
  Switzerland: IUCN. ix + 94pp. Available at the following web site:
  \url{iucnrle.org/}
}
\seealso{
Other gridUncertainty functions: \code{\link{gridUncertaintyBase}},
  \code{\link{gridUncertainty}}
}
\author{
Calvin Lee \email{calvinkflee@gmail.com}
}
