% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EOO_functions.R
\name{makeEOO}
\alias{makeEOO}
\title{Creates Extent of occurrence (EOO) Polygon}
\usage{
makeEOO(ecosystem.data)
}
\arguments{
\item{ecosystem.data}{Raster object of an ecosystem or species distribution.
Please use a CRS with units measured in metres.}
}
\value{
An object of class SpatialPolygons representing the EOO of
  \code{ecosystem.data}. Also inherits its CRS.
}
\description{
\code{makeEOO} creates a  minimum convex polygon enclosing all occurrences of
the provided data
}
\references{
Bland, L.M., Keith, D.A., Miller, R.M., Murray, N.J. and
  Rodriguez, J.P. (eds.) 2016. Guidelines for the application of IUCN Red
  List of Ecosystems Categories and Criteria, Version 1.0. Gland,
  Switzerland: IUCN. ix + 94pp. Available at the following web site:
  \url{iucnrle.org/}
}
\seealso{
Other EOO functions: \code{\link{getAreaEOO}}
}
\author{
Nicholas Murray \email{murr.nick@gmail.com}, Calvin Lee
  \email{calvinkflee@gmail.com}
}
