% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EOO_functions.R
\name{makeEOO}
\alias{makeEOO}
\title{Creates Extent of occurrence (EOO) Polygon}
\usage{
makeEOO(input.data, names_from)
}
\arguments{
\item{input.data}{Spatial object of an ecosystem or species distribution.
Please use a CRS with units measured in metres.}

\item{names_from}{name of the column containing ecosystem names.
If missing all features will be analysed together. Only needed for vector data.}
}
\value{
An object of class sf representing the EOO of
\code{input.data}, or a list of sf objects if multiple ecosystems were input.
Also inherits its CRS from input.data.
}
\description{
\code{makeEOO} is a generic function that creates a  minimum convex polygon
enclosing all occurrences of the ecosystems provided in the input data. If the input provided
is a raster layer, the points are taken from a buffer that has the radius of
half of the shorter edge of the pixel around the centroid.
}
\examples{
m <- matrix(sample(1:4, 500, replace = TRUE, prob = c(4,1,1,6)), nrow=25, ncol=20)
r1 <- terra::rast(m, crs = "EPSG:32755")
EOO.polygon <- makeEOO(r1)
}
\references{
Bland, L.M., Keith, D.A., Miller, R.M., Murray, N.J. and
Rodriguez, J.P. (eds.) 2016. Guidelines for the application of IUCN Red
List of Ecosystems Categories and Criteria, Version 1.0. Gland,
Switzerland: IUCN. ix + 94pp. Available at the following web site:
\url{https://iucnrle.org/}
}
\seealso{
Other EOO functions: 
\code{\link{getAreaEOO}()},
\code{\link{getEOO}()}
}
\author{
Nicholas Murray \email{murr.nick@gmail.com}, Calvin Lee
\email{calvinkflee@gmail.com}
}
\concept{EOO functions}
