% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{plot,AOOgrid,ANY-method}
\alias{plot,AOOgrid,ANY-method}
\title{AOOgrid plot}
\usage{
\S4method{plot}{AOOgrid,ANY}(x, title = x@name)
}
\arguments{
\item{x}{an AOOgrid object}

\item{title}{plot title, defaults to the assessment unit name.}
}
\description{
plot method for AOOgrid object
}
