% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{plot,AOOgrid,ANY-method}
\alias{plot,AOOgrid,ANY-method}
\title{AOOgrid plot}
\usage{
\S4method{plot}{AOOgrid,ANY}(x, y, ...)
}
\arguments{
\item{x}{an AOOgrid object}

\item{y}{Ignored. Included for consistency with the \code{plot} generic.}

\item{...}{Additional graphical parameters passed to \code{\link[graphics]{plot}}.}
}
\description{
plot method for AOOgrid object
}
