% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{plot,EOO,ANY-method}
\alias{plot,EOO,ANY-method}
\title{EOO plot}
\usage{
\S4method{plot}{EOO,ANY}(x, y, ...)
}
\arguments{
\item{x}{an EOO object}

\item{y}{Ignored. Included for consistency with the \code{plot} generic.}

\item{...}{Additional graphical parameters passed to \code{\link[graphics]{plot}}.}
}
\description{
plot method for EOO object
}
